lora: make the lora driver not dependent on lora node
The lora node radio driver should not be dependent on the lora node code. The radio
driver should exist as an independent entity. The changes made for this commit
are to make the driver stand-alone.

Note that part of this change removed some of the statistics. These were initial
stats to get the code up and running and were useful for the initial lora apps. Soon
there will be lora mfrg fw so not sure if the original apps are still useful. If they
are, the statistics should be added to the sx1276 code and should be added to the driver
as opposed to lora node.
